Move examples to Uncurried.re.
