+++ examples/typescript-react-example/src/ImportJsValue.bs.js
@@ -1,6 +1,5 @@
 // Generated by BUCKLESCRIPT VERSION 4.0.18, PLEASE EDIT WITH CARE
 
-import * as Curry from "bs-platform/lib/es6/curry.js";
 import * as ImportJsValueGen from "./ImportJsValue.gen";
 
 var roundedNumber = ImportJsValueGen.round(1.8);
@@ -24,22 +23,12 @@ function useGetAbs(x) {
   return x.getAbs() + 1 | 0;
 }
 
-function callback(cb) {
-  return String(Curry._1(cb, /* () */0));
-}
-
-function callback2(auth) {
-  return Curry._1(auth[/* login */0], /* () */0);
-}
-
 export {
   roundedNumber ,
   areaValue ,
   AbsoluteValue ,
   useGetProp ,
   useGetAbs ,
-  callback ,
-  callback2 ,
   
 }
 /* roundedNumber Not a pure module */
