+++ examples/typescript-react-example/src/Uncurried.bs.js
@@ -1,5 +1,6 @@
 // Generated by BUCKLESCRIPT VERSION 4.0.18, PLEASE EDIT WITH CARE
 
+import * as Curry from "bs-platform/lib/es6/curry.js";
 
 function uncurried0() {
   return "";
@@ -21,12 +22,22 @@ function curried3(x, y, z) {
   return String(x) + (y + String(z));
 }
 
+function callback(cb) {
+  return String(Curry._1(cb, /* () */0));
+}
+
+function callback2(auth) {
+  return Curry._1(auth[/* login */0], /* () */0);
+}
+
 export {
   uncurried0 ,
   uncurried1 ,
   uncurried2 ,
   uncurried3 ,
   curried3 ,
+  callback ,
+  callback2 ,
   
 }
 /* No side effect */
