+++ examples/typescript-react-example/src/Uncurried.gen.tsx
@@ -3,6 +3,9 @@
 // tslint:disable-next-line:no-var-requires
 const UncurriedBS = require('./Uncurried.bs');
 
+// tslint:disable-next-line:interface-over-type-literal
+export type auth = { readonly login: (_1:void) => string };
+
 export const uncurried0: () => string = UncurriedBS.uncurried0;
 
 export const uncurried1: (_1:number) => string = UncurriedBS.uncurried1;
@@ -12,3 +15,19 @@ export const uncurried2: (_1:number, _2:string) => string = UncurriedBS.uncurrie
 export const uncurried3: (_1:number, _2:string, _3:number) => string = UncurriedBS.uncurried3;
 
 export const curried3: (_1:number, _2:string, _3:number) => string = UncurriedBS.curried3;
+
+export const callback: (_1:((_1:void) => number)) => string = function _(Arg1: any) {
+  const result = UncurriedBS.callback(function _(Arg11: any) {
+      const result1 = Arg1(Arg11);
+      return result1
+    });
+  return result
+};
+
+export const callback2: (_1:auth) => string = function _(Arg1: any) {
+  const result = UncurriedBS.callback2([function _(Arg11: any) {
+      const result1 = Arg1.login(Arg11);
+      return result1
+    }]);
+  return result
+};
